backend/utils/webscraper/deployed/requirements.txt
annotated-types==0.7.0
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0

beautifulsoup4
python-dateutil

dateparser
selenium